refactor(prototyper): trap
`trap.rs` is complex enough to refactor.
First we can make all trap expcet boot use `fast_trap` crate,
second we can split this file into boot, handler and itself.
